Store the random order in the session. Closes #5019
