+++ decidim-budgets/app/controllers/decidim/budgets/projects_controller.rb
@@ -6,21 +6,18 @@ module Decidim
     class ProjectsController < Decidim::Budgets::ApplicationController
       include FilterResource
       include NeedsCurrentOrder
+      include Orderable
 
-      helper_method :projects, :random_seed, :project
+      helper_method :projects, :project
 
       private
 
       def projects
-        @projects ||= search.results.page(params[:page]).per(current_component.settings.projects_per_page)
-      end
-
-      def random_seed
-        @random_seed ||= search.random_seed
+        @projects ||= search.results.order_randomly(random_seed).page(params[:page]).per(current_component.settings.projects_per_page)
       end
 
       def project
-        @project ||= projects.find(params[:id])
+        @project ||= search.results.find(params[:id])
       end
 
       def search_klass
@@ -31,8 +28,7 @@ module Decidim
         {
           search_text: "",
           scope_id: "",
-          category_id: "",
-          random_seed: params[:random_seed]
+          category_id: ""
         }
       end
