+++ decidim-budgets/app/views/decidim/budgets/projects/_projects.html.erb
@@ -3,4 +3,4 @@
     <%= render partial: "project", locals: { project: project } %>
   <% end %>
 </div>
-<%= decidim_paginate projects, random_seed: random_seed %>
+<%= decidim_paginate projects %>
