+++ decidim-consultations/app/views/decidim/consultations/consultations/_consultations.html.erb
@@ -13,4 +13,4 @@
     <%= card_for consultation %>
   <% end %>
 </div>
-<%= decidim_paginate consultations, random_seed: random_seed %>
+<%= decidim_paginate consultations %>
