+++ decidim-sortitions/app/views/decidim/sortitions/sortitions/_sortitions.html.erb
@@ -11,4 +11,4 @@
 <div class="row small-up-1 medium-up-2 card-grid">
   <%= render @sortitions %>
 </div>
-<%= decidim_paginate @sortitions, random_seed: random_seed %>
+<%= decidim_paginate @sortitions %>
